requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
pydantic>=2.0.0
typing-extensions>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
tenacity>=8.0.0
PyYAML>=6.0.0